Create CitiesGraphExample
The content is similar to the adjacency list representation. Each row of the file defines the edges for a single vertex. For example, the first for of the file defines the edges of vertex A. In each row, after the "->" sign, edges are listed and separated with ",". Each edge definition consists of two elements, separated with ":". The first elements is the end-point and the second element is the weight of the edge. For example, the first edge that is defined for the vertex A is "B: 3". This means there is an edge from A to B of weight 3.